+++ crates/cli/src/bin/wasm-bindgen-test-runner/main.rs
@@ -16,7 +16,6 @@ use clap::Parser;
 use clap::ValueEnum;
 use std::env;
 use std::fs;
-use std::path::Path;
 use std::path::PathBuf;
 use std::thread;
 use wasm_bindgen_cli_support::Bindgen;
@@ -101,12 +100,6 @@ fn main() -> anyhow::Result<()> {
     let cli = Cli::parse();
 
     let shell = shell::Shell::new();
-
-    let file_name = cli
-        .file
-        .file_name()
-        .map(Path::new)
-        .context("file to test is not a valid file, can't extract file name")?;
 
     // Collect all tests that the test harness is supposed to run. We assume
     // that any exported function with the prefix `__wbg_test` is a test we need
@@ -276,8 +269,6 @@ fn main() -> anyhow::Result<()> {
         b.split_linked_modules(true);
     }
 
-    let coverage = coverage_args(file_name);
-
     b.debug(debug)
         .input_module(module, wasm)
         .keep_debug(false)
@@ -288,7 +279,7 @@ fn main() -> anyhow::Result<()> {
 
     match test_mode {
         TestMode::Node { no_modules } => {
-            node::execute(module, tmpdir.path(), cli, &tests, !no_modules, coverage)?
+            node::execute(module, tmpdir.path(), cli, &tests, !no_modules)?
         }
         TestMode::Deno => deno::execute(module, tmpdir.path(), cli, &tests)?,
         TestMode::Browser { .. }
@@ -310,7 +301,6 @@ fn main() -> anyhow::Result<()> {
                 &tests,
                 test_mode,
                 std::env::var("WASM_BINDGEN_TEST_NO_ORIGIN_ISOLATION").is_err(),
-                coverage,
             )
             .context("failed to spawn server")?;
             let addr = srv.server_addr();
@@ -376,28 +366,6 @@ impl TestMode {
             TestMode::SharedWorker { .. } => "WASM_BINDGEN_USE_SHARED_WORKER",
             TestMode::ServiceWorker { .. } => "WASM_BINDGEN_USE_SERVICE_WORKER",
         }
-    }
-}
-
-fn coverage_args(file_name: &Path) -> PathBuf {
-    fn generated(file_name: &Path, prefix: &str) -> String {
-        let res = format!("{prefix}{}.profraw", file_name.display());
-        res
-    }
-
-    let prefix = env::var_os("WASM_BINDGEN_UNSTABLE_TEST_PROFRAW_PREFIX")
-        .map(|s| s.to_str().unwrap().to_string())
-        .unwrap_or_default();
-
-    match env::var_os("WASM_BINDGEN_UNSTABLE_TEST_PROFRAW_OUT") {
-        Some(s) => {
-            let mut buf = PathBuf::from(s);
-            if buf.is_dir() {
-                buf.push(generated(file_name, &prefix));
-            }
-            buf
-        }
-        None => PathBuf::from(generated(file_name, &prefix)),
     }
 }